updated command's architecture
