+++ package.json
@@ -14,6 +14,7 @@
     "@strapi/plugin-i18n": "4.1.5",
     "@strapi/plugin-users-permissions": "4.1.5",
     "@strapi/strapi": "^4.1.5",
+    "axios": "^0.27.2",
     "eslint": "^8.12.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-config-standard": "^16.0.3",
